Usability improvements in StartTeamWindow
- Enlarge the window's width so a larger part of the path will fit in the TextBox.

- Apply the existing path to the BFilePanel before browsing. It improve
usability when you paste a path or type a new one in the TextBox
then click on the browse button to choose the application
or fine tune the directory.

Please not that it is a best effort feature : if the file or the directory does
not exist, the BFilePanel will fallback to the old behavior :
showing the previously selected file/directory in the session
(or home directory).

Change-Id: I504a1ec72f9f7a236e2e37039961fddc58218eae
Reviewed-on: https://review.haiku-os.org/672
Reviewed-by: Rene Gollent <rene@gollent.com>